style EditCalendar
