+++ src/components/editHatch/editHatch.css
@@ -1,0 +1,5 @@
+#dropdown-item-button {
+  background-color: #f9f5f3;
+  color: black;
+  border-color: #f9f5f3;
+}
